% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bern}
\alias{Bern}
\title{Sub-daily meteorological observations for Bern}
\format{
A list of data frames (one data frame per variable).
The format of the data frames is that required by the QC functions.
}
\source{
Institute of Geography - University of Bern
}
\usage{
Bern
}
\description{
Observations of pressure and temperature for the city of Bern (Switzerland)
for the period 1800-1827.
}
\keyword{datasets}
