% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{check_sef}
\alias{check_sef}
\title{Check compliance with SEF guidelines}
\usage{
check_sef(file = file.choose())
}
\arguments{
\item{file}{Character string giving the path of the SEF file.}
}
\value{
TRUE if no errors are found, FALSE otherwise.
}
\description{
Check compliance with SEF guidelines
}
\note{
For more information on error/warning messages produced by this
function see the SEF documentation.
}
\author{
Yuri Brugnara
}
