% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{climexp_to_sef}
\alias{climexp_to_sef}
\title{Download a GHCN-Daily data file from the Climate Explorer and convert it
into the Station Exchange Format version 0.2.0}
\usage{
climexp_to_sef(url, outpath = getwd())
}
\arguments{
\item{url}{Character string giving the url of the data file.}

\item{outpath}{Character string giving the path where to save the file.
By default this is the working directory.}
}
\description{
Download a GHCN-Daily data file from the Climate Explorer and convert it
into the Station Exchange Format version 0.2.0
}
\examples{
# Download the maximum temperature series for Buenos Aires
climexp_to_sef("http://climexp.knmi.nl/data/xgdcnAR000875850.dat")

}
\author{
Yuri Brugnara
}
