% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta}
\alias{Meta}
\title{Metadata for the stations of Bern and Rosario de Santa Fe}
\format{A list of data frames (one data frame per variable)}
\source{
Institute of Geography - University of Bern
}
\usage{
Meta
}
\description{
Metadata for the stations of Bern and Rosario de Santa Fe
}
\keyword{datasets}
