% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_meta}
\alias{read_meta}
\title{Read metadata from the Station Exchange Format version 0.2.0}
\usage{
read_meta(file = file.choose(), parameter = NULL)
}
\arguments{
\item{file}{Character string giving the path of the data file.}

\item{parameter}{Character vector of required parameters. Accepted
values are \code{"version"}, \code{"id"}, \code{"name"}, \code{"lat"},
\code{"lon"}, \code{"alt"}, \code{"source"}, \code{"link"},
\code{"var"}, \code{"stat"}, \code{"units"}, \code{"meta"}. 
By default all parameters are read at once.}
}
\value{
A character vector with the required parameters.
}
\description{
Read metadata from the Station Exchange Format version 0.2.0
}
\author{
Yuri Brugnara
}
