% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rosario}
\alias{Rosario}
\title{Daily and sub-daily meteorological observations for Rosario de Santa Fe (Argentina)}
\format{
A list of data frames (one data frame per variable).
The format of the data frames is that required by the QC functions.
}
\source{
ACRE
}
\usage{
Rosario
}
\description{
Observations of minimum and maximum temperature, wind direction, cloud cover,
pressure, precipitation, air temperature, wet bulb temperature, relative humidity,
dew point, and wind speed for the city of Rosario de Santa Fe (Argentina) for the 
period 1886-1900.
}
\keyword{datasets}
